Rename DiscordService to ClientService and add MessageService
